+++ src/Atata.Cli/GlobalSuppressions.cs
@@ -4,5 +4,6 @@
 
 [assembly: SuppressMessage("Design", "CA1000:Do not declare static members on generic types", Justification = "<Pending>", Scope = "member", Target = "~M:Atata.Cli.ProgramCli`1.InBaseDirectory~`0")]
 [assembly: SuppressMessage("Design", "CA1000:Do not declare static members on generic types", Justification = "<Pending>", Scope = "member", Target = "~M:Atata.Cli.ProgramCli`1.InDirectory(System.String)~`0")]
+[assembly: SuppressMessage("Style", "IDE0032:Use auto property", Justification = "<Pending>", Scope = "member", Target = "~F:Atata.Cli.CliCommand._process")]
 
 #pragma warning restore S103 // Lines should not be too long